+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/CommonAbstracts/HorizontallyAccured.kt
@@ -1,0 +1,7 @@
+package dev.inmo.tgbotapi.CommonAbstracts
+
+import dev.inmo.tgbotapi.types.Meters
+
+interface HorizontallyAccured {
+    val horizontalAccuracy: Meters?
+}