+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/CommonAbstracts/ProximityAlertable.kt
@@ -1,0 +1,7 @@
+package dev.inmo.tgbotapi.CommonAbstracts
+
+import dev.inmo.tgbotapi.types.Meters
+
+interface ProximityAlertable {
+    val proximityAlertRadius: Meters?
+}