+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/requests/edit/LiveLocation/EditInlineMessageLiveLocation.kt
@@ -3,6 +3,7 @@ package dev.inmo.tgbotapi.requests.edit.LiveLocation
 import dev.inmo.tgbotapi.requests.edit.abstracts.*
 import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.buttons.InlineKeyboardMarkup
+import dev.inmo.tgbotapi.utils.throwRangeError
 import kotlinx.serialization.*
 
 @Serializable
@@ -13,10 +14,22 @@ data class EditInlineMessageLiveLocation(
     override val latitude: Double,
     @SerialName(longitudeField)
     override val longitude: Double,
+    @SerialName(horizontalAccuracyField)
+    override val horizontalAccuracy: Meters? = null,
+    @SerialName(headingField)
+    override val heading: Degrees? = null,
+    @SerialName(proximityAlertRadiusField)
+    override val proximityAlertRadius: Meters? = null,
     @SerialName(replyMarkupField)
     override val replyMarkup: InlineKeyboardMarkup? = null
 ) : EditInlineMessage, EditReplyMessage, EditLocationMessage {
     override fun method(): String = "editMessageLiveLocation"
     override val requestSerializer: SerializationStrategy<*>
         get() = serializer()
+
+    init {
+        if (horizontalAccuracy != null && horizontalAccuracy !in horizontalAccuracyLimit) {
+            throwRangeError("horizontalAccuracy", horizontalAccuracyLimit, horizontalAccuracy)
+        }
+    }
 }