+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/requests/send/SendLocation.kt
@@ -1,19 +1,74 @@
 package dev.inmo.tgbotapi.requests.send
 
+import dev.inmo.tgbotapi.CommonAbstracts.*
 import dev.inmo.tgbotapi.requests.send.abstracts.*
 import dev.inmo.tgbotapi.types.*
 import dev.inmo.tgbotapi.types.buttons.KeyboardMarkup
 import dev.inmo.tgbotapi.types.message.abstracts.ContentMessage
 import dev.inmo.tgbotapi.types.message.abstracts.TelegramBotAPIMessageDeserializationStrategyClass
 import dev.inmo.tgbotapi.types.message.content.LocationContent
+import dev.inmo.tgbotapi.utils.throwRangeError
 import kotlinx.serialization.*
 
 
 private val commonResultDeserializer: DeserializationStrategy<ContentMessage<LocationContent>>
     = TelegramBotAPIMessageDeserializationStrategyClass()
 
+fun SendLocation(
+    chatId: ChatIdentifier,
+    latitude: Double,
+    longitude: Double,
+    disableNotification: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null,
+    replyMarkup: KeyboardMarkup? = null
+) = SendLocation(
+    chatId,
+    latitude,
+    longitude,
+    null,
+    null,
+    null,
+    null,
+    disableNotification,
+    replyToMessageId,
+    replyMarkup
+)
+
+fun SendStaticLocation(
+    chatId: ChatIdentifier,
+    latitude: Double,
+    longitude: Double,
+    disableNotification: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null,
+    replyMarkup: KeyboardMarkup? = null
+) = SendLocation(chatId, latitude, longitude, disableNotification, replyToMessageId, replyMarkup)
+
+fun SendLiveLocation(
+    chatId: ChatIdentifier,
+    latitude: Double,
+    longitude: Double,
+    livePeriod: Seconds,
+    horizontalAccuracy: Meters? = null,
+    heading: Degrees? = null,
+    proximityAlertRadius: Meters? = null,
+    disableNotification: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null,
+    replyMarkup: KeyboardMarkup? = null
+) = SendLocation(
+    chatId,
+    latitude,
+    longitude,
+    livePeriod,
+    horizontalAccuracy,
+    heading,
+    proximityAlertRadius,
+    disableNotification,
+    replyToMessageId,
+    replyMarkup
+)
+
 @Serializable
-data class SendLocation(
+data class SendLocation internal constructor(
     @SerialName(chatIdField)
     override val chatId: ChatIdentifier,
     @SerialName(latitudeField)
@@ -21,7 +76,13 @@ data class SendLocation(
     @SerialName(longitudeField)
     override val longitude: Double,
     @SerialName(livePeriodField)
-    val livePeriod: Long? = null,
+    override val livePeriod: Seconds? = null,
+    @SerialName(horizontalAccuracyField)
+    override val horizontalAccuracy: Meters? = null,
+    @SerialName(headingField)
+    override val heading: Degrees? = null,
+    @SerialName(proximityAlertRadiusField)
+    override val proximityAlertRadius: Meters? = null,
     @SerialName(disableNotificationField)
     override val disableNotification: Boolean = false,
     @SerialName(replyToMessageIdField)
@@ -30,7 +91,11 @@ data class SendLocation(
     override val replyMarkup: KeyboardMarkup? = null
 ) : SendMessageRequest<ContentMessage<LocationContent>>,
     ReplyingMarkupSendMessageRequest<ContentMessage<LocationContent>>,
-    PositionedSendMessageRequest<ContentMessage<LocationContent>>
+    PositionedSendMessageRequest<ContentMessage<LocationContent>>,
+    HorizontallyAccured,
+    Livable,
+    ProximityAlertable,
+    Headed
 {
     override fun method(): String = "sendLocation"
     override val resultDeserializer: DeserializationStrategy<ContentMessage<LocationContent>>
@@ -41,6 +106,9 @@ data class SendLocation(
     init {
         if (livePeriod != null && livePeriod !in livePeriodLimit) {
             error("Live period for sending location must be in $livePeriodLimit, but was $livePeriod")
+        }
+        if (horizontalAccuracy != null && horizontalAccuracy !in horizontalAccuracyLimit) {
+            throwRangeError("horizontalAccuracy", horizontalAccuracyLimit, horizontalAccuracy)
         }
     }
 }