+++ tgbotapi.extensions.api/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/api/edit/LiveLocation/EditInlineMessageLiveLocation.kt
@@ -3,7 +3,7 @@ package dev.inmo.tgbotapi.extensions.api.edit.LiveLocation
 import dev.inmo.tgbotapi.bot.TelegramBot
 import dev.inmo.tgbotapi.requests.edit.LiveLocation.EditInlineMessageLiveLocation
 import dev.inmo.tgbotapi.types.InlineMessageIdentifier
-import dev.inmo.tgbotapi.types.Location
+import dev.inmo.tgbotapi.types.location.StaticLocation
 import dev.inmo.tgbotapi.types.buttons.InlineKeyboardMarkup
 
 suspend fun TelegramBot.editLiveLocation(
@@ -18,6 +18,6 @@ suspend fun TelegramBot.editLiveLocation(
 )
 suspend fun TelegramBot.editLiveLocation(
     inlineMessageId: InlineMessageIdentifier,
-    location: Location,
+    location: StaticLocation,
     replyMarkup: InlineKeyboardMarkup? = null
 ) = editLiveLocation(inlineMessageId, location.latitude, location.longitude, replyMarkup)